+++ src/black/mode.py
@@ -264,6 +264,7 @@ class Preview(Enum):
     hug_comparator = auto()
     parenthesize_tuple_in_yield = auto()
     fmt_off_class_blank_lines = auto()
+    remove_redundant_generator_parentheses = auto()
 
 
 UNSTABLE_FEATURES: set[Preview] = {